+++ src/components/Editor/Runners/PythonRunner/PythonRunner.js
@@ -6,6 +6,8 @@ import { useSelector, useDispatch } from "react-redux";
 import { useTranslation } from "react-i18next";
 import { Tab, Tabs, TabList, TabPanel } from "react-tabs";
 import Sk from "skulpt";
+import { useMediaQuery } from "react-responsive";
+
 import {
   setError,
   codeRunHandled,
@@ -14,11 +16,12 @@ import {
   triggerDraw,
 } from "../../EditorSlice";
 import ErrorMessage from "../../ErrorMessage/ErrorMessage";
-
 import store from "../../../../app/store";
 import VisualOutputPane from "./VisualOutputPane";
 import OutputViewToggle from "./OutputViewToggle";
 import { SettingsContext } from "../../../../settings";
+import RunnerControls from "../../../RunButton/RunnerControls";
+import { MOBILE_MEDIA_QUERY } from "../../../../utils/mediaQueryBreakpoints";
 
 const externalLibraries = {
   "./pygal/__init__.js": {
@@ -66,6 +69,7 @@ const PythonRunner = () => {
   const dispatch = useDispatch();
   const { t } = useTranslation();
   const settings = useContext(SettingsContext);
+  const isMobile = useMediaQuery({ query: MOBILE_MEDIA_QUERY });
 
   const queryParams = new URLSearchParams(window.location.search);
   const [hasVisualOutput, setHasVisualOutput] = useState(
@@ -373,6 +377,7 @@ const PythonRunner = () => {
                     </Tab>
                   </TabList>
                   {!isEmbedded ? <OutputViewToggle /> : null}
+                  {!isEmbedded && isMobile ? <RunnerControls skinny /> : null}
                 </div>
                 <TabPanel key={0}>
                   <VisualOutputPane />
@@ -391,6 +396,9 @@ const PythonRunner = () => {
                   </Tab>
                 </TabList>
                 {hasVisualOutput || isEmbedded ? null : <OutputViewToggle />}
+                {!hasVisualOutput && !isEmbedded && isMobile ? (
+                  <RunnerControls skinny />
+                ) : null}
               </div>
               <ErrorMessage />
               <TabPanel key={0}>
@@ -421,6 +429,7 @@ const PythonRunner = () => {
               </Tab>
             </TabList>
             {!isEmbedded ? <OutputViewToggle /> : null}
+            {!isEmbedded && isMobile ? <RunnerControls skinny /> : null}
           </div>
           <ErrorMessage />
           {hasVisualOutput ? (